Tests: add verify_route_not_registered failure case, refactoring
